Modified the live session to correctly save pending updates'
